backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
mysqlclient>=2.1
Pillow>=10.0
django-filter>=23.0
drf-yasg>=1.21
